Revert unpacking change due to bug
The url package task defines it's schema using 2-tuples as it has no interest
in the "forward to ckan" option. This therefore caused an issue when attempting
to unpack a 2-tuple whilst expecting a 3-tuple. This is resolved in this commit
by reverting back to the previous code which doesn't bother to unpack the schema.