Make detecting the gbif zip outputs easier for the scripts
We automatically rebuild the gbif output of the data portal each week. It is,
however, hard to determine which zips are for gbif and which ones are for normal
users. We used to use the existance of an eml.xml file to do it but we now include
that file in all dwc downloads so we can't do that. This is a bit hacky but it's
the safest way to determine which zip is a gbif zip.