Increase celery version to 3.1.25
This is so that we can support python versions >= 2.7.11. Celery 2.7.20 contains the actual bump in version to kombu we need (kombu >= 3.3.30). This came about because they were using a private method from a builtin module: https://github.com/celery/kombu/issues/545. Celery 3.1.25 has this fix and is the highest 3.1.x version available so might as well go up to that.