Support XLSX files as a download format
Only available to datastore downloads that aren't in DwC-A.
This is implemented in a pretty hacky way - we create a CSV file and then convert it to XLSX in the finalise step. The reason for doing it this way is to cause the least disruption possible as this is legacy code and will be completely deprecated soon. Therefore, it is better to not disrupt the current code with a larger change but instead to just tack on an extra bonus bit.